Move DeviceInfo and DeviceVerification to separate module
